+++ src/Application/VinylRecordComments/Commands/DeleteVinylRecordCommentCommand.cs
@@ -1,17 +1,20 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using Application.Common.Errors;
 using Application.Common.Interfaces.Repositories;
+using ErrorOr;
 using MediatR;
+using Optional;
 
 namespace Application.VinylRecordComments.Commands
 {
-    public record DeleteVinylRecordCommentCommand : IRequest
+    public record DeleteVinylRecordCommentCommand : IRequest<ErrorOr<Deleted>>
     {
         public required Guid Id { get; init; }
     }
 
-    public class DeleteVinylRecordCommentCommandHandler : IRequestHandler<DeleteVinylRecordCommentCommand>
+    public class DeleteVinylRecordCommentCommandHandler : IRequestHandler<DeleteVinylRecordCommentCommand, ErrorOr<Deleted>>
     {
         private readonly IVinylRecordCommentRepository _commentRepository;
 
@@ -20,9 +23,16 @@ namespace Application.VinylRecordComments.Commands
             _commentRepository = commentRepository;
         }
 
-        public async Task Handle(DeleteVinylRecordCommentCommand request, CancellationToken cancellationToken)
+        public async Task<ErrorOr<Deleted>> Handle(DeleteVinylRecordCommentCommand request, CancellationToken cancellationToken)
         {
+            var existingOption = await _commentRepository.GetByIdAsync(request.Id, cancellationToken);
+            if (!existingOption.HasValue)
+            {
+                return Errors.VinylRecordComment.NotFound(request.Id);
+            }
+
             await _commentRepository.DeleteAsync(request.Id, cancellationToken);
+            return Result.Deleted;
         }
     }
 }
